pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "productivity-agent"
version = "0.1.5"
description = "A suite of scripts that connect your Google Workspace activity to Google Tasks."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Rhys" },
]
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.3.1",
    "google-auth-httplib2>=0.1.0",
    "python-dotenv>=1.2.2",
    "python-dateutil>=2.8.2",
    "yt-dlp>=2026.3.17",
    "youtube-transcript-api>=1.2.4",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pylint",
]

[tool.pylint.TYPECHECK]
generated-members = ["googleapiclient.*"]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["agent_utils"]
script-files = [
    "src/prod_agent.py",
    "src/prod_agent_meet.py",
    "src/prod_agent_tasks.py",
    "src/prod_agent_notebooklm.py",
    "src/prod_agent_podcast.py",
    "src/setup_auth.py",
    "scripts/security-scan.py",
]
